Add create-local-package script to generate local copy of package for development
This is a better approach than just installing the actual components folder directly, since we get a properly packaged version with all dependencies properly resolved to use in development. Installing the folder directly tends to link the entire package repository to node_modules, which is not quite what we want. We only want package.json and dist to appear in node_modules. This approach of generating a .tgz file and installing that ensures this.