Add FileInput components from Dropcore
There's not really any need for a separate package for those files - it just adds one extra package to maintain when there's already so many.

I also do want to change the implementation slightly. I'm not sure how necessary the mapping of file types to file mimes are - we could potentially just use the extensions directly.