Standardise error message formatting
The pattern I generally go for with errors is to use all capitals with words separated by underscore when throwing an error message. This standardised code can then get translated into a more human-readable error in any UI in which the error may occur. This makes the error code easier to check and compare against, while still allowing for us to display a more readable one later if needed.