Add Loader component
This gets rid of the need to keep doing if(isLoading){ return <Loading/> } on every single page in Neurosongs. Now we can just wrap the page around a Loader and pass in our loading state. This makes it much easier to handle loading states and can work with any querying framework that provides some sort of loading state for the query.

Note that while Suspense is a thing already, it only works with specific frameworks, and I don't think @tanstack/react-query is one of them. As such, this is still worth keeping around as a framework-agnostic wrapper to handle loading states.