Add LoaderProvider component
This is nice to use if you want the same convenience of loading/error state handling that the base Loader gives you, but you also want more control over exactly where you display the error and loading components on the page. To use it, you would surround your page in a LoaderProvider, passing it the loading state and error state, and then use LoaderContent for content you want to display when the page is not loading, and LoaderError to display any errors.

LoaderContent also takes a showOnError prop, allowing us to choose whether to show the content if an error has been thrown or not. Hiding is the default behaviour so that we avoid potential side effects, but we can opt into showing on error by using showOnError, which we may use if we know that there are certain parts of the UI that we can still render even if the query fails.