Refactor base Loader component to use the new LoaderProvider and related components
This allows users to keep the previous implementation where we can just surround a part of the UI with the Loader component without needing to worry about setting up an entire context. It's just more that the context helps in cases where you want more customisation, but if not, the usual Loader works well.