Add LoaderData component
This uses the data from the LoaderProvider and runs some checks to ensure that it is not undefined when isLoading is false. If it is, there is a chance to handle that with the onUndefined prop, otherwise we can display the output of the children function evaluated with the data.

Note that for LoaderData usage, you do have to pass in the type to LoaderData in order for the type inference to properly work. Without it, the data will be typed as unknown, which is not often what we want. We can still trust that the data will not be undefined when we run the function, though - it's just a case of types not really being able to be passed down the context due to type-checking being done at compile time. If you do want the nice type inferencing without needing to pass the type, you will have to use the Loader component instead, which isn't quite as flexible but still works well for most use cases.