Add onNull and onNullable handlers
This gives us a nice way to deal with cases where the data is null without having to deal with it in the children. This is especially nice for the source code for LoaderData now because it means that we don't have to do that type casting when passing into children anymore. The return types can truly be NonNullable because the component now natively supports the edge cases for data being null.