Simplify the Loader types
It pretty much has everything that LoaderProviderProps has and everything LoaderDataProps has, except for the fact that the children type from LoaderData should overrule, and we don't really need showOnError because Loader always shows an error if the error state is provided and is defined.

Note that a ts-expect-error is needed to suppress the LoaderData mutual exclusivity type checks in Loader, as Loader has to pass them all through separately. Since it specified all of them, the discriminated union is violated, however, since I know for a fact that Loader is also enforcing the same mutual exclusivity rules, we can disable the type errors here.