Make LoaderError the one responsible for displaying undefined/null/nullable components
LoaderError is really the one that's meant to deal with edge cases like this, whereas LoaderData should ideally be primarily focused on handling when things go right. As such, I've decided to move the nullable handlers to LoaderError and also make them ReactNode types rather than functions. This makes it clearer that they should ideally be more concerned with displaying a message on the UI in the event of this happening.