Deprecate usage of nullable handlers on LoaderData
LoaderData is really meant to take care of the best-case scenario where you have all the data you need, whereas LoaderError is what takes care of the worst-case scenario. Letting LoaderData take the handlers feels like it muddies the boundary a bit, and I feel like it'd be better to let LoaderError handle it while LoaderData only checks if the data exists or not.