Update dependencies and address eslint-plugin-react-hooks errors
For the refs errors, I just ended up disabling it as I'm fairly sure my implementation in LoaderError is correct. Adhering to the rule would most likely require setting up some useEffects, which would overcomplicate the logic, and given that it's mainly being used to determine whether we should log to the console, I think that seems like a common enough use case for useRef, so I think it might be best to disable that rule.

Also note that the useEffect in ScreenSizeProvider has been replaced with a useMemo. I think, the general better practice in React whenever you want to useEffect + setState is to instead try and opt for useMemo, as that more or less combines the two - it returns a stateful value based on a function that runs based on the dependency array. Also note that the largeScreenCondition helper no longer needs to be here as it was only being used in the useMemo anyway, so we can just put the contents there directly.