Add script to update peer dependencies
We have just been in a situation where we had version conflicts due to peer dependencies being very outdated, as well as a very fiddly mismatch between typescript-eslint packages. As such, I think it's best to try and update peer dependencies at the start of every month. It's probably not best to update them as regularly as the others, as other projects rely on peer dependencies being relatively stable, but they should be kept fairly up-to-date.