Add --no-frozen-lockfile to update-dependencies script
Apparently, PNPM uses --frozen-lockfile in the GitHub Actions by default. This is a smart idea, to be fair, as it ensures that the CI environment doesn't unintentionally change the lockfile, but for the update action to run, we need to change the lockfile. As such, we need to add the --no-frozen-lockfile flag so that the update-dependencies action can actually update.