Get rid of npm-check-updates reliance
pnpm update is actually a lot more nicer than npm update. npm update never really worked as it should've as it didn't change the package.json file with the updated versions, so I had to rely on npm-check-updates to have the package.json versions changed. pnpm update, on the other hand, actually changes the version numbers in package.json as well as the lockfile, which is the behaviour we want. As such, we no longer need another dependency to control dependency updates (which feels silly to me) and we can now just rely natively on the package manager to do it for us. That is a massive win right here.