Use the update-dependencies script
This allows us to use the --latest flag to ensure that we do update to major versions, while still allowing us to specify which dependencies do not get updated. That is, we can say that tsdown must not be updated, and it will always get ignored even if you were to pass it into the script. And then, when we are ready to update tsdown, we can just remove the negation.