Switch the update command to be pnpm update --latest && pnpm update
Essentially what this says is, we first try updating all dependencies to latest. This is, of course, not always the best idea as there may be some dependencies we know definitely won't work with latest. As such, we have the second step of pnpm update, which falls back to what we do know works based on the overrides. If there's no override, we stay on the guaranteed latest version (which may even be a new major version), but if there is, we fall back to the override.

Once the updates have been made, a pull request goes up testing the updates work as intended. if CI passes, we can assume that the new versions are ok to accept and continue with them. If it does not pass, we must either fix the issue to allow us to be up-to-date with the latest version, or specify an override for the time being and fix it some other time.